backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.94",
    "fastapi>=0.136.0",
    "mangum>=0.21.0",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "uvicorn>=0.45.0",
]
